requirements.txt
uvicorn[standard]==0.23.2
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.146
botocore==1.34.146
Brotli==1.1.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cohere==5.6.2
cssselect2==0.7.0
dataclasses-json==0.6.7
distro==1.9.0
fastavro==1.9.5
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
html2text==2024.2.26
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.1
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.1.13
langchain-cohere==0.1.5
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.1
langchain-text-splitters==0.0.2
langgraph==0.0.32
langsmith==0.1.93
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.37.0
orjson==3.10.6
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
parameterized==0.9.0
pdfkit==1.0.0
pillow==10.4.0
protobuf==5.27.2
pyarrow==17.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.18.0
PyMuPDF==1.24.8
PyMuPDFb==1.24.8
pyphen==0.15.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
reportlab==4.2.2
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
s3transfer==0.10.2
selenium==4.23.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
streamlit==1.36.0
tenacity==8.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
weasyprint==62.3
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
zopfli==0.2.3